backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.18
pdfplumber==0.11.4
PyMuPDF==1.24.14
python-docx==1.1.2
openpyxl==3.1.5
markdown==3.7
redis==5.2.0
aiofiles==24.1.0
python-magic==0.4.27
reportlab==4.2.5
Pillow==11.0.0
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.24.0
